Add prototype of downloader.
Write a tool for downloading the data of a URL for display to standard
output as text.  Binary files are converted to a Base-64 form.  MIME
header information describing the content type precedes the data.
Include imported third-party libraries for the Swift Package Manager.

Refine which files generated by the Swift Package Manager should be
excluded by Git.